+++ example/lib/shared/decoration/chest.dart
@@ -111,9 +111,5 @@ class Chest extends GameDecoration with TapGesture {
     );
   }
 
-  @override
-  void onTapDown(int pointer, Vector2 position) {}
 
-  @override
-  void onTapUp(int pointer, Vector2 position) {}
 }