+++ example/lib/shared/npc/wizard/wizard.dart
@@ -108,13 +108,4 @@ class Wizard extends SimpleNpc with ObjectCollision, TapGesture {
   void onTap() {
     execShowTalk(this);
   }
-
-  @override
-  void onTapCancel() {}
-
-  @override
-  void onTapDown(int pointer, Vector2 position) {}
-
-  @override
-  void onTapUp(int pointer, Vector2 position) {}
 }